Allow compile-time enabling/disabling indrev
