+++ build.gradle
@@ -105,11 +105,13 @@ dependencies {
     }
 
     // Industrial Revolution
-    dependencies.ext.mod "curse.maven:industrial-revolution-391708:${project.industrial_rev_id}"
-    dependencies.ext.mod "net.fabricmc:fabric-language-kotlin:${project.kotlin_version}"
-    dependencies.ext.mod "io.github.cottonmc:LibGui:${project.libgui_version}"
-    dependencies.ext.mod "dev.cafeteria:fake-player-api:${project.fake_player_version}"
-    dependencies.ext.mod "com.github.GabrielOlvH:magna:${project.magna_version}"
+    if (indrev_enabled.toBoolean()) {
+        dependencies.ext.mod "curse.maven:industrial-revolution-391708:${project.industrial_rev_id}"
+        dependencies.ext.mod "net.fabricmc:fabric-language-kotlin:${project.kotlin_version}"
+        dependencies.ext.mod "io.github.cottonmc:LibGui:${project.libgui_version}"
+        dependencies.ext.mod "dev.cafeteria:fake-player-api:${project.fake_player_version}"
+        dependencies.ext.mod "com.github.GabrielOlvH:magna:${project.magna_version}"
+    }
 
     //
     // Stuff we just want to have at runtime
@@ -124,6 +126,9 @@ sourceSets {
         java {
             if (!tech_reborn_enabled.toBoolean()) {
                 exclude('com/kneelawk/extramodintegrations/techreborn')
+            }
+            if (!indrev_enabled.toBoolean()) {
+                exclude('com/kneelawk/extramodintegrations/industrialrevolution')
             }
         }
     }
