+++ gradle.properties
@@ -43,6 +43,7 @@ reborn_core_id = 4102771
 tech_reborn_id = 4102770
 
 # Industrial Revolution
+indrev_enabled = true
 libgui_version = 6.4.0+1.19
 industrial_rev_id = 4020957
 kotlin_version = 1.8.7+kotlin.1.7.22
